refactor(agent/runtime): extract coordinator infrastructure into coordinators/ subdirectory
The agent/runtime directory contained 45+ files covering unrelated concerns
with no internal structure. The five promise-based coordinator files
(BasePromiseCoordinator, PlanApprovalCoordinator, RetryRequestCoordinator,
AgentProposalCoordinator, runCoordinators) form a cohesive subsystem that
gates on user UI decisions (retry, plan approval, proposal), but were buried
alongside session management, registry, lifecycle, model factory, and helper
utilities with no grouping signal.

Moving them to coordinators/ makes the pattern immediately discoverable,
reduces the cognitive load of navigating runtime/, and establishes a precedent
for further sub-module extraction (session/, registry/, etc.).

No logic changes — pure file relocation with import path updates across 21
call sites in src/, packages/cli/, packages/extension/, and test-kernel/.
