+++ src/agent/runtime/SessionHandle.ts
@@ -36,7 +36,7 @@ import { createChannelTrace } from '@logger';
 import { tryUseRunContext } from './RunContext';
 import { ExecutionRegistry, executionRegistry } from './executionRegistry';
 import { InterruptRegistry, interruptRegistry } from './InterruptRegistry';
-import { RunCoordinatorBridge, runCoordinatorBridge } from './runCoordinators';
+import { RunCoordinatorBridge, runCoordinatorBridge } from './coordinators/runCoordinators';
 import {
   ExecutionSubscriptionBinder,
   executionSubscriptionBinder,
