+++ src/agent/runtime/coordinators/AgentProposalCoordinator.ts
