+++ src/agent/runtime/coordinators/runCoordinators.ts
@@ -1,5 +1,5 @@
-import { useRunContext, type RunCoordinators } from './RunContext';
-import { executionRegistry, type ExecutionRegistry } from './executionRegistry';
+import { useRunContext, type RunCoordinators } from '../RunContext';
+import { executionRegistry, type ExecutionRegistry } from '../executionRegistry';
 import type {
   ProposalRequestOptions,
   ProposalResult,
