+++ src/controllers/progressView/ProgressAgentProposalController.ts
@@ -4,7 +4,7 @@ import {
   type AgentConfig,
 } from '@agent/core/definition/AgentConfig';
 import { AgentCategory } from '@agent/core/definition/AgentDataclass';
-import type { ProposalResult } from '@agent/runtime/AgentProposalCoordinator';
+import type { ProposalResult } from '@agent/runtime/coordinators/AgentProposalCoordinator';
 
 import type { TaskState } from '@agent/core/execution/TaskState';
