+++ src/test-kernel/agent/runtime/runCoordinators.vitest.ts
@@ -4,15 +4,15 @@ import { describe, expect, it, vi } from 'vitest';
 // Local imports - runtime
 import type { AgentTrace } from '@agent/trace';
 import type { AgentRuntimeHost } from '@agent/runtime/AgentRuntimeHost';
-import { AgentProposalCoordinator } from '@agent/runtime/AgentProposalCoordinator';
-import { PlanApprovalCoordinator } from '@agent/runtime/PlanApprovalCoordinator';
-import { RetryRequestCoordinatorImpl } from '@agent/runtime/RetryRequestCoordinator';
+import { AgentProposalCoordinator } from '@agent/runtime/coordinators/AgentProposalCoordinator';
+import { PlanApprovalCoordinator } from '@agent/runtime/coordinators/PlanApprovalCoordinator';
+import { RetryRequestCoordinatorImpl } from '@agent/runtime/coordinators/RetryRequestCoordinator';
 import {
   createRunContext,
   withRunContext,
   type RunCoordinators,
 } from '@agent/runtime/RunContext';
-import { RunCoordinatorBridge } from '@agent/runtime/runCoordinators';
+import { RunCoordinatorBridge } from '@agent/runtime/coordinators/runCoordinators';
 import {
   AgentExecutionHandle,
   ExecutionRegistry,
